subclass OccupancyGridMapBase
