+++ perception/probabilistic_occupancy_grid_map/config/pointcloud_based_occupancy_grid_map.param.yaml
@@ -18,10 +18,13 @@
     # center of the grid map
     gridmap_origin_frame: "base_link"
     # ray-tracing center: main sensor frame is preferable like: "velodyne_top"
+    # base_link should not be used with "OccupancyGridMapProjectiveBlindSpot"
     scan_origin_frame: "base_link"
 
-    use_projection: false
-    projection_dz_threshold: 0.01 # [m] for avoiding null division
-    obstacle_separation_threshold: 1.0 # [m] fill the interval between obstacles with unknown for this length
-
-    pub_debug_grid: false
+    grid_map_type: "OccupancyGridMapFixedBlindSpot"
+    OccupancyGridMapFixedBlindSpot:
+      distance_margin: 1.0
+    OccupancyGridMapProjectiveBlindSpot:
+      projection_dz_threshold: 0.01 # [m] for avoiding null division
+      obstacle_separation_threshold: 1.0 # [m] fill the interval between obstacles with unknown for this length
+      pub_debug_grid: false
