+++ perception/probabilistic_occupancy_grid_map/pointcloud-based-occupancy-grid-map.md
@@ -40,7 +40,7 @@ The ray trace is done by Bresenham's line algorithm.
    - The obstacle point cloud is processed and may not match the raw pointcloud.
    - The input may be inaccurate and obstacle points may not be determined as obstacles.
 
-   When the parameter `use_projeciton` is true, for each obstacle pointcloud, if there are no _visible_ raw pointclouds that are located above the projected ray from the `scan_origin` to that obstacle pointcloud, the cells between the obstacle pointcloud and the `projected point` are filled with `UNKNOWN`. Note that the `scan_origin` should not be `base_link` if this flag is true because otherwise all the cells behind the obstacle point clouds would be filled with `UNKNOWN`.
+   When the parameter `grid_map_type` is "OccupancyGridMapProjectiveBlindSpot" and the `scan_origin` is a sensor frame like `velodyne_top` for instance, for each obstacle pointcloud, if there are no _visible_ raw pointclouds that are located above the projected ray from the `scan_origin` to that obstacle pointcloud, the cells between the obstacle pointcloud and the `projected point` are filled with `UNKNOWN`. Note that the `scan_origin` should not be `base_link` if this flag is true because otherwise all the cells behind the obstacle point clouds would be filled with `UNKNOWN`.
 
    ![pointcloud_based_occupancy_grid_map_side_view_2nd_projection](./image/pointcloud_based_occupancy_grid_map_side_view_2nd_projection.drawio.svg)
 
@@ -88,6 +88,7 @@ $$
 | `use_height_filter` | bool   | whether to height filter for `~/input/obstacle_pointcloud` and `~/input/raw_pointcloud`? By default, the height is set to -1~2m. |
 | `map_length`        | double | The length of the map. -100 if it is 50~50[m]                                                                                    |
 | `map_resolution`    | double | The map cell resolution [m]                                                                                                      |
+| `grid_map_type`     | string | The type of grid map for estimating `UNKNOWN` region behind obstacle point clouds                                                |
 
 ## Assumptions / Known limits
 
@@ -110,8 +111,14 @@ In several places we have modified the external code written in BSD3 license.
 
 ## How to debug
 
-set `pub_debug_grid` to `true` and
+If `grid_map_type` is "OccupancyGridMapProjectiveBlindSpot" and `pub_debug_grid` is `true`, it is possible to check the each process of grid map generation by running
 
 ```shell
-ros2 launch autoware_launch logging_simulator.launch.xml vehicle_model:=sample_vehicle sensor_model:=sample_sensor_kit map_path:=<sample-map-rosbag> rviz_config:=$(ros2 pkg prefix --share probabilistic_occupancy_grid_map)/config/debug.rviz planning:=false control:=false api:=false system:=false
+ros2 launch probabilistic_occupancy_grid_map debug.launch.xml
 ```
+
+and visualizing the following occupancy grid map topics (which are listed in config/grid_map_param.yaml):
+
+- `/perception/occupancy_grid_map/grid_1st_step`: `FREE` cells are filled
+- `/perception/occupancy_grid_map/grid_2nd_step`: `UNKNOWN` cells are filled
+- `/perception/occupancy_grid_map/grid_3rd_step`: `OCCUPIED` cells are filled
